Report AgentCard, remove uneeded imports, use A2A Types
- Let a2a library handle JSON-RPC errors as designed
- Remove custom jsonrpc_error_handler function and related imports
- Move AgentCard out of routes to keep it to only endpoint handlers
