+++ src/agent/api/__init__.py
@@ -6,7 +6,6 @@ from .app import app, create_app, main
 from .routes import (
     create_agent_card,
     get_request_handler,
-    jsonrpc_error_handler,
     router,
     set_request_handler_instance,
     sse_generator,
@@ -18,7 +17,6 @@ __all__ = [
     "main",
     "create_agent_card",
     "get_request_handler",
-    "jsonrpc_error_handler",
     "router",
     "set_request_handler_instance",
     "sse_generator",
