+++ src/agent/config/a2a.py
@@ -7,9 +7,6 @@ for JSON-RPC communication between agents.
 
 from __future__ import annotations
 
-from abc import ABC
-from typing import Any
-
 # Import official A2A types
 from a2a.types import (
     AgentCapabilities,
@@ -34,15 +31,6 @@ from a2a.types import (
     TaskStatus,
     TextPart,
 )
-from pydantic import BaseModel, Field
-
-
-class JSONRPCError(Exception):
-    def __init__(self, code: int, message: str, data: Any = None):
-        super().__init__(message)
-        self.code = code
-        self.message = message
-        self.data = data
 
 
 class TaskNotFoundError(Exception):
@@ -92,44 +80,6 @@ def get_error_code_for_exception(exception_type: type[Exception]) -> int | None:
     return A2A_ERROR_CODE_MAP.get(exception_type)
 
 
-def create_jsonrpc_error_from_exception(exception: Exception, request_id: Any = None) -> dict[str, Any]:
-    """Create a JSON-RPC error response from an exception.
-
-    Args:
-        exception: The exception instance
-        request_id: The JSON-RPC request ID
-
-    Returns:
-        A JSON-RPC error response dictionary
-    """
-    error_code = get_error_code_for_exception(type(exception))
-
-    if error_code is None:
-        # Default to internal error for unknown exceptions
-        error_code = -32603
-
-    return {
-        "jsonrpc": "2.0",
-        "error": {"code": error_code, "message": str(exception), "data": {"exception_type": type(exception).__name__}},
-        "id": request_id,
-    }
-
-
-# NOTE: Configuration models have been moved to model.py
-# This file now focuses solely on A2A protocol integration
-
-
-class BaseAgent(BaseModel, ABC):
-    model_config = {
-        "arbitrary_types_allowed": True,
-        "extra": "allow",
-    }
-
-    agent_name: str = Field(description="The name of the agent.")
-    description: str = Field(description="A brief description of the agent's purpose.")
-    content_types: list[str] = Field(description="Supported content types.")
-
-
 # Re-export A2A types and error handling for convenience
 __all__ = [
     # A2A protocol types
@@ -155,7 +105,6 @@ __all__ = [
     "TaskState",
     "TaskStatus",
     # A2A JSON-RPC exceptions
-    "JSONRPCError",
     "TaskNotFoundError",
     "TaskNotCancelableError",
     "PushNotificationNotSupportedError",
@@ -165,7 +114,4 @@ __all__ = [
     # A2A error handling utilities
     "A2A_ERROR_CODE_MAP",
     "get_error_code_for_exception",
-    "create_jsonrpc_error_from_exception",
-    # A2A base classes
-    "BaseAgent",
 ]
